add more models

+++ sde/models/inv_positions.py
@@ -1,10 +1,11 @@
 from django.db import models
 
+
 class Position(models.Model):
     item_id = models.IntegerField(primary_key=True)
-    pitch = models.DecimalField()
-    roll = models.DecimalField()
-    x = models.DecimalField()
-    y = models.DecimalField()
-    z = models.DecimalField()
-    yaw = models.DecimalField()
+    pitch = models.FloatField()
+    roll = models.FloatField()
+    x = models.FloatField()
+    y = models.FloatField()
+    z = models.FloatField()
+    yaw = models.FloatField()
